move existing screens to legacy foldercan delete later. just don't want to delete them entirely for right now.
